Popular ingredients and recipes with missing
List popular ingredients that you do not have in your shelf. This is a
quick way of adding ingredients you are likely to own but have not yet
added to your shelf.

Add recipes with more than one missing ingredient to UI3. This is useful
when you are focusing in on a single ingredient as you are much more
likely to match no recipes.